Extract MainFunctionDetectorBase without dependency on PSI
The main motivation for this change is to be able to use
MainFunctionDetectorBase in modules which don't/shouldn't have access to
PSI, such as compiler/ir/serialization.jvm, which is why it's been
extracted to module 'resolution'.

The most non-trivial part of MainFunctionDetectorBase is the logic of
obtaining "main" candidates from the same file as another candidate. It
doesn't seem possible to perform it effectively only on descriptors, so
there's a new abstract method getFunctionsFromTheSameFile which is
supposed to provide this behavior (in MainFunctionDetector, it still
works as before, via PSI).